Fix visit node when buffer is already narrowed.
## Changes Description
Before this commit, if the buffer was narrowed when you try to visit a
header, the visit would fail.

This commit widens thee buffer before vising the section, and also
introduces a new behavior for the universal prefix argument. Now, when a
buffer is narrowed, selecting a section with the prefix argument will
toggle the narrow state. This allows the user to manipulate the
narrowed state from the outline itself.