Fix broken build
## Changes Description

Last commits introduce a break on the unit tests build due to not
requiring `all-the-icons` as a mandatory library. This commit changes
the usage of `all-the-icon` to be dependent on the availability of thee
library.

Also fixed the use of `cl-defstruct` and add the missing default values
to the slots, and a pretty bad case of _copy & paste_ when I renamed the
`org-ol-tree/visit-section` on commit `6dcafc4`.