Add auto-resize for the outline window
## Changes Description

This commit introduces a feature that I wanted to implement since the
inception of this project: The auto-resize!

The idea is that your outline window should occupy as little size as
possible, but it should also show as much of its content as possible.
The problem I had in the past was to achieve this when the window uses a
variable-pitch face to display its contents (like what I want to use in
my personal config). On this commit I borrow some Emacs code from the
`fit-window-to-buffer` and was able to create a feature-rich auto-resize
configuration.

The user can control if auto-resize is enable (which it is by default),
as well as the max size (in pixels, columns, or frame percent), and a
flag to tell org-ol-tree that it should auto-resize its window only when
the new size is bigger then the current one.