Small changes on current tests
## Changes Description

This commit renames the section test file to the more generic name
`test-outline-elements.el` to match test files with code sections.

It also enhance one test to verify that empty headings have their level
set to the integer 0, and also to check that the subheadings slot is an
empty list.