Use `fboundp` instead of `require` with `'noerror`
## Changes Description

The commit 6dcafc4 added the basis for the full key binding, and that
feature tries to understand if Evil is available so it can also provide
Evil bindings.

Unfortunately, on some Emacs installations (I believe this happens only
on Emacs 28.0.50), the use of `require` with the `'noerror` option is
returning a non `nil` value even if the feature is not available and
can't be loaded.

This commit tries to work around the problem by checking with `fboundp`
if all the functions necessary for Evil integration and `all-the-icons`
as well, are present before trying to use those features.

Fix #10.