Fix last line highlight and update README
## Changes Description

Treemacs builds its buffer and finishes in the line of the last item.
The effect of this, is the line highlight of last line to be shorter
than the other lines.

This commit adds a '\n' at the end of last line, causing the line
highlight to fully highlight the last element.