Fix broken build
## Changes Description

Previous commit, added more tests to the project, but it didn't properly
cleanup these tests. The result was that `default-directory` was set to
the directory of a data file used in the tests when all tests finished.
At that point, undercover created the coverage file on
`default-directory`, which was not the expected location, causing the
build to break.

This commit fix this issue.