Customize the rename child frame to fix UI issues
## Changes Description

I've got feedback from users that the renaming functionality, although
working, has some UX issues:

- Child frame is unaligned with the window;
- Can't exit the child frame;
- After exiting renaming, there are modifications on the Org file;
- After finishing renaming and adjusting the Org buffer, there is a gap
  on the tree;

I believe most (if not all) of these issues, are related to the library
`cfrs` that Treemacs uses (and I decided to use on my package as well).
Unfortunately, this library does not allow me to do the customization I
need in order to address the reported concerns, so this commit, is an
attempt to do just that.